# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mongo Micro Service API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://csprd-mongo.crazysales.com.au/v0/"
        }
    ],
    "paths": {
        "/version": {
            "get": {
                "tags": [
                    "version"
                ],
                "summary": "Get framework version",
                "operationId": "App\\Http\\Controllers\\ExampleController::version",
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/productlog/{productID}?skip={skip}&take={take}": {
            "get": {
                "tags": [
                    "product log"
                ],
                "summary": "Get a product log data by productID",
                "operationId": "findByProductID",
                "parameters": [
                    {
                        "name": "productID",
                        "in": "path",
                        "description": "Such as: 123",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "Numbers ,default is 1",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "take",
                        "in": "query",
                        "description": "Numbers ,default is all",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product log data"
                    }
                }
            }
        },
        "/productlog/create": {
            "post": {
                "tags": [
                    "product log"
                ],
                "summary": "Add a new product log data",
                "description": "",
                "operationId": "App\\Http\\Controllers\\ProductlogController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "admin": {
                                        "type": "string"
                                    },
                                    "brief": {
                                        "type": "string"
                                    },
                                    "details": {
                                        "type": "string"
                                    },
                                    "productID": {
                                        "type": "string"
                                    },
                                    "remark": {
                                        "type": "string"
                                    },
                                    "date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "return inserted data"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}